app/requirements.txt
beautifulsoup4==4.4.1
boardgamegeek==0.13.2
Django==1.9.4
django-appconf==1.0.1
django-compressor==2.0
django-extensions==1.6.1
django-modelcluster==1.1
django-taggit==0.18.0
django-treebeard==4.0
djangorestframework==3.3.3
html5lib==0.9999999
jsonfield==1.0.3
Pillow==3.1.1
pytz==2016.3
rcssmin==1.0.6
requests==2.9.1
requests-cache==0.4.10
rjsmin==1.0.12
six==1.10.0
sorl-thumbnail==12.4a1
Unidecode==0.4.19
wagtail==1.5.2
Willow==0.3
psycopg2==2.6.1
